telegram_bot/requirements.txt
aiogram>=3.22.0
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.54.0
faster-whisper>=1.1.0
fastapi>=0.110.0
uvicorn>=0.27.0
PyJWT>=2.8.0
python-multipart>=0.0.6
